Create new multiindex_object.py
